Separate to two Validation tabs in sample ui plugin
Separate to two Validation tabs in sample ui plugin
