+++ packages/ui-plugins/sample-ui-plugin/src/components/Validation.jsx
@@ -17,16 +17,9 @@
 
 import React from 'react';
 import { ValidationItem } from '@assurance/validation-summary';
-import { Button, ComboBox, defaultTheme, Provider, Item} from '@adobe/react-spectrum';
-import { useState, useEffect } from 'react';
-import { Flex } from '@adobe/react-spectrum';
-import { 
-  useValidation,
-} from '@assurance/plugin-bridge-provider';
-
 
 //directly add validatoin items to the validation view
-/*const namespaces = ['aep-consent-configuration', 'aep-consent-registered', 'adobe-core-configuration'];
+const namespaces = ['aep-consent-configuration', 'aep-consent-registered', 'adobe-core-configuration'];
 
 const Validation = () => {
 return (
@@ -35,40 +28,5 @@ return (
      </>
    );
  };
- */
-
-//add validation items to the validation view by selecting from a dropdown
-const Validation = () => {
-  const [selected, setSelected] = useState('');
-  const [selectedItems, setSelectedItems] = useState([]);
-  const fullSetValidation = useValidation();
-  const [namespaces, setNamespaces] = useState([])
-
-useEffect(() => {
-  if (fullSetValidation) {
-    setNamespaces(Object.keys(fullSetValidation));
-  }
-}, [fullSetValidation]);
-
-const handleAddValidation = () => {
-  if (selected && !selectedItems.includes(selected)) {
-    setSelectedItems([...selectedItems, selected]);
-  }
-};
-
-return (  
-  <Provider theme={defaultTheme} colorScheme="light">
-    <Flex direction="row" gap="size-500" alignItems="end">
-      <ComboBox label="Choose a validation" selectedKey={selected} onSelectionChange={setSelected}>
-        {namespaces.map(namespace => (
-          <Item key={namespace}>{namespace}</Item>
-        ))}
-      </ComboBox>
-      <Button variant="cta" onPress={handleAddValidation}>Add Validation</Button>
-      </Flex>
-      {selectedItems.map(item => <ValidationItem key={item} namespace={item} />)}
-    </Provider>
-  );
-};
 
 export default Validation;
